:white_check_mark: Fixed mock setup for tests
