+++ src/components/DataDisplay/OptionalTooltip.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { render, screen } from '../../test-utils';
+import { render, screen } from '../../tests/test-utils';
 import OptionalTooltip from './OptionalTooltip';
 
 const dummyData = {
