+++ src/components/DataDisplay/ProfileAvatar.test.tsx
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { render, screen } from '../../test-utils';
+import { render, screen } from '../../tests/test-utils';
 import ProfileAvatar, { ProfileAvatarProps } from './ProfileAvatar';
 
 function mockProfileAvatarProps(image: boolean): ProfileAvatarProps {
