+++ src/components/DataDisplay/Table/Table.test.tsx
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { render, screen } from '../../../test-utils';
+import { render, screen } from '../../../tests/test-utils';
 import Table from './Table';
 import { getIcon, TableItemProps } from './TableItem';
