+++ src/components/DataDisplay/Tutorial/Tutorial.test.tsx
@@ -1,7 +1,7 @@
 import { faker } from '@faker-js/faker';
 
 import TutorialStepsProvider from '../../../providers/TutorialStepsProvider';
-import { render, screen, userEvent } from '../../../test-utils';
+import { render, screen, userEvent } from '../../../tests/test-utils';
 import Tutorial, { IStep, TutorialProps } from './Tutorial';
 
 function fakeStep(): IStep {
