+++ src/components/Feedback/FullPageSpinner.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { tokens } from '@equinor/eds-tokens';
 
-import { render, screen } from '../../test-utils';
+import { render, screen } from '../../tests/test-utils';
 import FullPageSpinner from './FullPageSpinner';
 
 const { colors } = tokens;
