+++ src/components/Feedback/FullPageUnauthorized.test.tsx
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { render, screen } from '../../test-utils';
+import { render, screen } from '../../tests/test-utils';
 import FullPageUnauthorized from './FullPageUnauthorized';
 
 test('Shows appName prop correctly', () => {
