+++ src/components/Feedback/NotFound.test.tsx
@@ -2,7 +2,7 @@ import { Icon } from '@equinor/eds-core-react';
 import { car_wash } from '@equinor/eds-icons';
 import { faker } from '@faker-js/faker';
 
-import { render, screen, userEvent } from '../../test-utils';
+import { render, screen, userEvent } from '../../tests/test-utils';
 import NotFound from './NotFound';
 
 function fakeProps() {
