+++ src/components/Feedback/SelectField/SelectField.test.tsx
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { render, screen, userEvent, waitFor } from '../../../test-utils';
+import { render, screen, userEvent, waitFor } from '../../../tests/test-utils';
 import { Field } from '../../../types/Field';
 import SelectField from './SelectField';
