+++ src/components/Feedback/Stepper/Stepper.test.tsx
@@ -1,7 +1,7 @@
 import { check } from '@equinor/eds-icons';
 import { faker } from '@faker-js/faker';
 
-import { render, screen, userEvent } from '../../../test-utils';
+import { render, screen, userEvent } from '../../../tests/test-utils';
 import Stepper, { StepperProps } from './Stepper';
 
 function fakeProps(): StepperProps {
