+++ src/components/Inputs/ContentMenu.test.tsx
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { render, screen, userEvent } from '../../test-utils';
+import { render, screen, userEvent } from '../../tests/test-utils';
 import ContentMenu, { ContentMenuProps } from './ContentMenu';
 
 function fakeItem(): { label: string; value: string } {
