+++ src/components/Inputs/Select/SingleSelectDrawer.test.tsx
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { render, screen, userEvent } from '../../../test-utils';
+import { render, screen, userEvent } from '../../../tests/test-utils';
 import SingleSelectDrawer, {
   SingleSelectDrawerProps,
 } from './SingleSelectDrawer';
