+++ src/components/Inputs/Sieve/Sieve.test.tsx
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { render, screen, userEvent } from '../../../test-utils';
+import { render, screen, userEvent } from '../../../tests/test-utils';
 import { FilterOption } from './Filter';
 import Sieve, { SieveProps } from './Sieve';
 import { Option } from './Sieve.common';
