+++ src/components/Inputs/SingleFilterMenu.test.tsx
@@ -1,7 +1,7 @@
 import { filter_list } from '@equinor/eds-icons';
 import { faker } from '@faker-js/faker';
 
-import { render, screen, userEvent } from '../../test-utils';
+import { render, screen, userEvent } from '../../tests/test-utils';
 import SingleFilterMenu from './SingleFilterMenu';
 
 function getTestProps(): {
