+++ src/components/Inputs/TextEntry/EditableTextEntry.test.tsx
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { render, screen, userEvent } from '../../../test-utils';
+import { render, screen, userEvent } from '../../../tests/test-utils';
 import EditableTextEntry, { EditableTextEntryProps } from './EditableTextEntry';
 
 function fakeProps(): EditableTextEntryProps {
