+++ src/components/Inputs/TextEntry/TextEntry.test.tsx
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { render, screen, userEvent } from '../../../test-utils';
+import { render, screen, userEvent } from '../../../tests/test-utils';
 import TextEntry, { TextEntryProps } from './TextEntry';
 
 function fakeProps(): TextEntryProps {
