+++ src/components/Navigation/TopBar/Guidelines.test.tsx
@@ -7,7 +7,7 @@ import {
 } from '@equinor/eds-icons';
 import { faker } from '@faker-js/faker';
 
-import { render, screen } from '../../../test-utils';
+import { render, screen } from '../../../tests/test-utils';
 import { Guidelines, IGuidelineProps, IGuidelineSections } from './Guidelines';
 
 function fakeSection(): IGuidelineSections {
