+++ src/components/Navigation/TopBar/Settings.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen, userEvent } from '../../../test-utils';
+import { render, screen, userEvent } from '../../../tests/test-utils';
 import Settings, { ISettingsProps } from './Settings';
 
 test('Settings renders as expected', async () => {
