+++ src/components/Navigation/TopBar/TopBar.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { car } from '@equinor/eds-icons';
 
-import { render, screen } from '../../../test-utils';
+import { render, screen } from '../../tests/test-utils';
 import TopBar from '.';
 
 test('Shows progress indicator only when isFetching={true}', () => {
