+++ src/tests/mockResizeObserver.ts
@@ -1,0 +1,4 @@
+import { vi } from 'vitest';
+
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+vi.stubGlobal('ResizeObserver', require('resize-observer-polyfill'));
